+++ example/basic.go
@@ -16,11 +16,17 @@ func BasicEma() techan.Indicator {
 	// fetch this from your preferred exchange
 	dataset := [][]string{
 		// Timestamp, Open, Close, High, Low, volume
-		{"1234567", "1", "2", "3", "5", "6"},
+		{"1609459200", "99", "100", "101", "98", "12"},
+		{"1609545600", "100", "102", "103", "99", "15"},
+		{"1609632000", "102", "101", "104", "100", "11"},
+		{"1609718400", "101", "104", "105", "100", "18"},
 	}
 
 	for _, datum := range dataset {
-		start, _ := strconv.ParseInt(datum[0], 10, 64)
+		start, err := strconv.ParseInt(datum[0], 10, 64)
+		if err != nil {
+			panic(err)
+		}
 		period := techan.NewTimePeriod(time.Unix(start, 0), time.Hour*24)
 
 		candle := techan.NewCandle(period)
@@ -28,12 +34,15 @@ func BasicEma() techan.Indicator {
 		candle.ClosePrice = big.NewFromString(datum[2])
 		candle.MaxPrice = big.NewFromString(datum[3])
 		candle.MinPrice = big.NewFromString(datum[4])
+		candle.Volume = big.NewFromString(datum[5])
 
-		series.AddCandle(candle)
+		if !series.AddCandle(candle) {
+			panic("candles must be in chronological order")
+		}
 	}
 
 	closePrices := techan.NewClosePriceIndicator(series)
-	movingAverage := techan.NewEMAIndicator(closePrices, 10) // Create an exponential moving average with a window of 10
+	movingAverage := techan.NewEMAIndicator(closePrices, 3) // Three observations per window
 
 	return movingAverage
 }