+++ rule_position.go
@@ -9,7 +9,7 @@ func (pnr PositionNewRule) IsSatisfied(index int, record *TradingRecord) bool {
 	return record.CurrentPosition().IsNew()
 }
 
-//PositionOpenRule is satisfied when the current position in the trading record is open (position
+// PositionOpenRule is satisfied when the current position in the trading record is open (position
 // has been entered but not exited).
 type PositionOpenRule struct{}